Create README.md
Imagine witnessing an accident on a road, the people take the victim to a doctor, even after proper treatment the patient's condition isn't improving, infact it's getting worse. It turns out that the patient is allergic to the medicines administered by the doctor. There was no way for the doctor to know about this allergic condition of the patient. In modern India, with the push towards digitalization, we too need to upgrade to an online platform and giveup the use of paper base prescriptions. We plan to Implement a centralized database for healthcare where the doctors can record patient details and the patients can access it online. This will improve the accessibility of the records and can be a deciding factor in life and death situations.